Move User event listeners to Core
